Añadido expansión overworld + arreglo para compilar
editado los archivos para soportar mas overworlds y además, arreglados muchas cosas para que pudiese compilar el juego.

+++ include/union_room.h
@@ -96,7 +96,8 @@ struct WirelessLink_Group
 struct UnionRoomObject
 {
     u8 state;
-    u8 gfxId;
+    //u8 gfxId;
+    u16 gfxId;
     s8 animState;
     u8 schedAnim;
 };
